refactor: move pages from its folders
