+++ src/pages/Error404.tsx
